# app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Ubademy - Backend",
    "description": "Ubademy -  Coursera like app, for taller2",
    "license": {
      "name": "FIUBA",
      "url": "https://fi.uba.ar/"
    }
  },
  "host": "ubademy-la-venganza-back.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "first_name": {
                  "type": "string"
                },
                "last_name": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "birth_date": {
                  "type": "string",
                  "format": "date"
                },
                "phone_type": {
                  "type": "string"
                },
                "phone_number": {
                  "type": "string"
                },
                "email": {
                  "type": "string",
                  "format": "email"
                }
              }
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Create new user",
        "responses": {
          "201": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/user/{user_id}": {
      "parameters": [
        {
          "in": "path",
          "name": "user_id",
          "type": "string",
          "required": true,
          "description": "ID of the user"
        },
        {
          "in": "query",
          "name": "properties",
          "type": "string",
          "required": false
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by id",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "patch": {
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "first_name": {
                  "type": "string"
                },
                "last_name": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "birth_date": {
                  "type": "string",
                  "format": "date"
                },
                "phone_type": {
                  "type": "string"
                },
                "phone_number": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Modify user's content by id",
        "responses": {
          "202": {
            "description": "User inormation updated correctly"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/user/login/{email}": {
      "parameters": [
        {
          "in": "path",
          "name": "email",
          "type": "string",
          "required": true,
          "description": "Email of the user to get"
        },
        {
          "in": "query",
          "name": "properties",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user's content by email",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/user/admin/{email}": {
      "parameters": [
        {
          "in": "path",
          "name": "email",
          "type": "string",
          "required": true,
          "description": "Email of the admin user to get"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get admin user content by email",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/user/all": {
      "parameters": [
        {
          "in": "query",
          "name": "max_results",
          "type": "integer"
        },
        {
          "in": "query",
          "name": "keyword",
          "type": "string"
        },
        {
          "in": "query",
          "name": "page",
          "type": "integer"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get users",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
...
    "/exam/{exam_id}/course/{course_id}/lesson/{lesson_id}/user/{user_id}": {
      "get": {
        "tags": [
          "Exams"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "type": "integer",
            "required": true,
            "description": "ID for the exam to get"
          },
          {
            "in": "path",
            "name": "course_id",
            "type": "integer",
            "required": true,
            "description": "ID for the course to get exam from"
          },
          {
            "in": "path",
            "name": "lesson_id",
            "type": "integer",
            "required": true,
            "description": "ID for the lesson to get exam from"
          },
          {
            "in": "path",
            "name": "user_id",
            "type": "string",
            "required": true,
            "description": "ID for the user that requests the exam"
          }
        ],
        "summary": "Get exam's content by id",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
...
    "/wallet/{user_id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "type": "string",
            "required": true,
            "description": "ID of the user requesting their wallet address"
          }
        ],
        "tags": [
          "Wallet"
        ],
        "summary": "Get wallet address of a user",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/wallet/balance/{user_id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "type": "string",
            "required": true,
            "description": "ID of the user requesting the balance in their wallet"
          }
        ],
        "tags": [
          "Wallet"
        ],
        "summary": "Get wallet address of a user",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/wallet/teacherBalance/{user_id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "type": "string",
            "required": true,
            "description": "ID of the user requesting their payout balance"
          }
        ],
        "tags": [
          "Wallet"
        ],
        "summary": "Get balance of a teacher",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/wallet/teacherWithdraw/{user_id}": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "type": "string",
            "required": true,
            "description": "ID of the user requesting to withraw their payments"
          },
          {
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "recieverAddress"
              ],
              "properties": {
                "recieverAddress": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "Wallet"
        ],
        "summary": "Payout to wallet address given",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/wallet/admin/balance": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Wallet"
        ],
        "summary": "Get admin's balance",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/wallet/admin/withdraw": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "amount"
              ],
              "properties": {
                "amount": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "Wallet"
        ],
        "summary": "Get wallet address of a user",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}